Updated the NaCl(100) slab generator (#158)
In this PR, we update the slab with the cell parameters obtained with the UZH basis set and GTH potential containing one valence electron for Na. 9 valence electrons would require a cutoff of 1600 Ry. For the moment, we keep NaCl slabs with this setting. This might change in the future.