Fix issue with with exporting PDOS curves (#213)
Before this change, if one tries to press on the "TXT" link, one only gets
a file with the energies column. In the PR I fix such a behaviour and create
a PDOS file with both: energies and densities.

Also, since the output format of the file is effectively `CSV`, I changed the filename
from `pdos.txt` to `pdos.csv`.